backend/requirements.txt
fastapi==0.109.0
groq==0.4.2
pydantic==2.5.3
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
